requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0
greenlet>=3.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# CORS
aiofiles>=23.2.0

# Image handling
Pillow>=10.0.0
python-slugify>=8.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
